Do not use home directory for spooldirs but user IDs
Because there is a big problem with the current thing: multiple users
may have the same home directory and so multiple users may get the
same fasl root. That is a bad thing[tm] especially because this case
won't let act multiple users on that FASL root.

This breaks usage of clisp in sudo, because in sudo, my UID will be 0
but my home direcotry is still /home/rvb and I will get the fasl cache
of rvb as root. Bad bad bad ... This situation always arises when one
does "sudo aptitude install clisp" and causes trouble for a very cool
patch that I am about to send you as soon as this one is applied. ;-)

User IDs are unique, while directory names are not. I highly recommend
applying this patch. :-> Of course the current practise is more
portable, but not practicable. :-/

darcs-hash:20060109002440-006ab-5182613048f590525167d643fab02a10edf68d9b.gz