customizable images for ecl-alikes
To make Peter happy, the files the user wants to have added are
returned in a list, so that ECL-images may also be customized!

And finally, no more problems if the recreation is run with sudo,
because the fasl-root, that is fixedly generated for bootstrapping is
used and only for root.

That still doesn't change the fact that the current behaviour is
buggy. ;-)

darcs-hash:20060110165710-006ab-769ba65e70f84bd2a2fc323339a273280e9841ea.gz