make use of ECL's ASDF addons for customized images
ECL extends ASDF with a build-op operation that compiles an ASDF
system to .o files and links them to a .fasl or shared library. This
is used for load-component now so that .o files are built that can be
linked into the image.

Users will want to use these ASDF extensions, so we somehow have to
figure out how to get ECLs own ASDF (that is a fas) added to the image
instead of the "plain" one, or to extend the plain one with the
asdf-ecl.lisp file that is provided with ECL.

For now, I explicitly added a (require "asdf") for ECL in
post-sysdef-install.lisp.

darcs-hash:20060112154043-006ab-83b462adf6fc3ec0206338c5d1f1f03494822867.gz